+++ repos/zimeg/CHANGELOG.md
@@ -8,6 +8,8 @@ and are noted with a `!` in the message.
 
 ## Changes
 
+- ci(fix): attempt to build packages on all systems in spite of fails 2024-06-28
+- feat(pkgs/gon): package the program that signs and notarizes darwin 2024-06-28
 - build(macos): perform cache builds on both macos and ubuntu runners 2024-06-27
 - build(cachix): compile and cache built binaries for faster installs 2024-06-27
 - docs: instruct installation of private packages into project flakes 2024-06-27